requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
protobuf==4.21.12
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette
typing_extensions==4.4.0
uvicorn==0.20.0